make sure we have very very basic test scaffolding for all the new pages
