+++ plugins/plugin-site/src/templates/plugin_dependencies.test.jsx
@@ -1,0 +1,34 @@
+
+import React from 'react';
+
+import {render} from '@testing-library/react';
+
+import TemplatePluginDependencies from './plugin_dependencies';
+
+describe('component - TemplatePluginDependencies', () => {
+    it('renders', async () => {
+        const plugin = {
+            name: 'fake',
+            title: 'fake plugin',
+            buildDate: '0',
+            labels: [],
+            developers: [],
+            stats: {
+                installations: [],
+                currentInstalls: 15
+            },
+            gav: 'something',
+            dependencies: [],
+            wiki: {
+                url: 'https://somewhere.com',
+                content: 'something'
+            }
+        };
+        const {container} = render(<TemplatePluginDependencies data={{
+            jenkinsPlugin: plugin,
+            reverseDependencies: {edges: []}
+        }} />);
+
+        expect(container).toBeTruthy();
+    });
+});
